refactor: split `viewer.py` into multiple files
